pyproject.toml
[tool.poetry]
name = "qq-chat-bot"
version = "0.1.0"
description = "A QQ bot based on chatAI."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = "^2.0.0"
requests = "^2.28.0"
httpx = "^0.23.0" # nonebot需要

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"